+++ cmd/forklift/plt/cli.go
@@ -222,15 +222,28 @@ func makeQuerySubcmds() []*cli.Command {
 		makeQueryDeplSubcmds(category),
 		[]*cli.Command{
 			{
+				Name:     "ls-dl",
+				Aliases:  []string{"list-downloads"},
+				Category: category,
+				Usage:    "Lists the files to be downloaded for export by the local pallet",
+				Action:   lsDlAction,
+				Flags: []cli.Flag{
+					&cli.BoolFlag{
+						Name:  "include-disabled",
+						Usage: "Also list images for disabled package deployments",
+					},
+				},
+			},
+			{
 				Name:     "ls-img",
 				Aliases:  []string{"list-images"},
 				Category: category,
-				Usage:    "Lists the Docker container images required by the development pallet",
+				Usage:    "Lists the Docker container images required by the local pallet",
 				Action:   lsImgAction,
 				Flags: []cli.Flag{
 					&cli.BoolFlag{
 						Name:  "include-disabled",
-						Usage: "Also download images for disabled package deployments",
+						Usage: "Also list images for disabled package deployments",
 					},
 				},
 			},
@@ -305,7 +318,7 @@ func makeQueryFileSubcmds(category string) []*cli.Command {
 	return []*cli.Command{
 		{
 			Name:      "ls-file",
-			Aliases:   []string{"list-file", "ls-files", "list-files"},
+			Aliases:   []string{"list-files"},
 			Category:  category,
 			Usage:     "Lists non-directory files in the local pallet",
 			ArgsUsage: "[path_glob]",